cleanups before table element init redo
