+++ internal/wasm/engine.go
@@ -3,8 +3,14 @@ package internalwasm
 // Engine is a Store-scoped mechanism to compile functions declared or imported by a module.
 // This is a top-level type implemented by an interpreter or JIT compiler.
 type Engine interface {
-	// Compile compiles down the function instances in a module, and returns ModuleEngine for the module.
-	Compile(importedFunctions, moduleFunctions []*FunctionInstance) (ModuleEngine, error)
+	// NewModuleEngine compiles down the function instances in a module, and returns ModuleEngine for the module.
+	//
+	// * importedFunctions: functions this module imports, already compiled in this engine.
+	// * moduleFunctions: functions declared in this module that must be compiled.
+	//
+	// Note: Input parameters must be pre-validated with internalwasm.Module Validate, to ensure no fields are invalid
+	// due to reasons such as out-of-bounds.
+	NewModuleEngine(importedFunctions, moduleFunctions []*FunctionInstance) (ModuleEngine, error)
 }
 
 // ModuleEngine implements function calls for a given module.
@@ -17,7 +23,7 @@ type ModuleEngine interface {
 
 	// Call invokes a function instance f with given parameters.
 	// Returns the results from the function.
-	// The ctx's context.Context will be the outer-most ancestor of the argument to wasm.FunctionVoidReturn, etc.
+	// The ctx's context.Context will be the outer-most ancestor of the argument to wasm.Function.
 	Call(ctx *ModuleContext, f *FunctionInstance, params ...uint64) (results []uint64, err error)
 
 	// Release releases the resources allocated by functions in this ModuleEngine.
