+++ internal/wasm/jit/jit_arm64.go
@@ -1,5 +1,3 @@
-//go:build arm64
-
 // This file implements the compiler for arm64 target.
 // Please refer to https://developer.arm.com/documentation/102374/latest/
 // if unfamiliar with arm64 instructions and semantics.
