+++ internal/wasm/jit/jit_value_location_amd64.go
@@ -1,5 +1,3 @@
-//go:build amd64
-
 package jit
 
 import (
