+++ internal/wasm/store_test.go
@@ -439,7 +439,7 @@ func newStore() *Store {
 	return NewStore(context.Background(), &mockEngine{shouldCompileFail: false, callFailIndex: -1}, Features20191205)
 }
 
-func (e *mockEngine) Compile(_, _ []*FunctionInstance) (ModuleEngine, error) {
+func (e *mockEngine) NewModuleEngine(_, _ []*FunctionInstance) (ModuleEngine, error) {
 	if e.shouldCompileFail {
 		return nil, fmt.Errorf("some compilation error")
 	}
@@ -672,42 +672,6 @@ func TestStore_resolveImports(t *testing.T) {
 			require.EqualError(t, err, "import[0] global[test.target]: value type mismatch: f64 != i32")
 		})
 	})
-	t.Run("table", func(t *testing.T) {
-		t.Run("ok", func(t *testing.T) {
-			s := newStore()
-			max := uint32(10)
-			tableInst := &TableInstance{Max: &max}
-			s.modules[moduleName] = &ModuleInstance{Exports: map[string]*ExportInstance{name: {
-				Type:  ExternTypeTable,
-				Table: tableInst,
-			}}, Name: moduleName}
-			_, _, table, _, _, err := s.resolveImports(&Module{ImportSection: []*Import{{Module: moduleName, Name: name, Type: ExternTypeTable, DescTable: &Table{Max: &max}}}})
-			require.NoError(t, err)
-			require.Equal(t, table, tableInst)
-			require.Equal(t, 1, s.modules[moduleName].dependentCount)
-		})
-		t.Run("minimum size mismatch", func(t *testing.T) {
-			s := newStore()
-			importTableType := &Table{Min: 2}
-			s.modules[moduleName] = &ModuleInstance{Exports: map[string]*ExportInstance{name: {
-				Type:  ExternTypeTable,
-				Table: &TableInstance{Min: importTableType.Min - 1},
-			}}, Name: moduleName}
-			_, _, _, _, _, err := s.resolveImports(&Module{ImportSection: []*Import{{Module: moduleName, Name: name, Type: ExternTypeTable, DescTable: importTableType}}})
-			require.EqualError(t, err, "import[0] table[test.target]: minimum size mismatch: 2 > 1")
-		})
-		t.Run("maximum size mismatch", func(t *testing.T) {
-			s := newStore()
-			max := uint32(10)
-			importTableType := &Table{Max: &max}
-			s.modules[moduleName] = &ModuleInstance{Exports: map[string]*ExportInstance{name: {
-				Type:  ExternTypeTable,
-				Table: &TableInstance{Min: importTableType.Min - 1},
-			}}, Name: moduleName}
-			_, _, _, _, _, err := s.resolveImports(&Module{ImportSection: []*Import{{Module: moduleName, Name: name, Type: ExternTypeTable, DescTable: importTableType}}})
-			require.EqualError(t, err, "import[0] table[test.target]: maximum size mismatch: 10, but actual has no max")
-		})
-	})
 	t.Run("memory", func(t *testing.T) {
 		t.Run("ok", func(t *testing.T) {
 			s := newStore()
@@ -794,75 +758,6 @@ func TestModuleInstance_applyData(t *testing.T) {
 		{OffsetExpression: &ConstantExpression{Opcode: OpcodeI32Const, Data: []byte{0x8}}, Init: []byte{0x1, 0x5}},
 	})
 	require.Equal(t, []byte{0xa, 0xf, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0, 0x1, 0x5}, m.Memory.Buffer)
-}
-
-func TestModuleInstance_validateElements(t *testing.T) {
-	functionCounts := uint32(0xa)
-	m := &ModuleInstance{
-		Table:     &TableInstance{Table: make([]uintptr, 10)},
-		Functions: make([]*FunctionInstance, 10),
-	}
-	for _, tc := range []struct {
-		name     string
-		elements []*ElementSegment
-		expErr   bool
-	}{
-		{
-			name: "ok",
-			elements: []*ElementSegment{
-				{OffsetExpr: &ConstantExpression{Opcode: OpcodeI32Const, Data: []byte{0x0}}, Init: []uint32{0, functionCounts - 1}},
-			},
-		},
-		{
-			name: "ok on edge",
-			elements: []*ElementSegment{
-				{OffsetExpr: &ConstantExpression{Opcode: OpcodeI32Const, Data: []byte{0x8}}, Init: []uint32{0, functionCounts - 1}},
-			},
-		},
-		{
-			name: "out of bounds",
-			elements: []*ElementSegment{
-				{OffsetExpr: &ConstantExpression{Opcode: OpcodeI32Const, Data: []byte{0x9}}, Init: []uint32{0, functionCounts - 1}},
-			},
-			expErr: true,
-		},
-		{
-			name: "unknown function",
-			elements: []*ElementSegment{
-				{OffsetExpr: &ConstantExpression{Opcode: OpcodeI32Const, Data: []byte{0x0}}, Init: []uint32{0, functionCounts}},
-			},
-			expErr: true,
-		},
-	} {
-		tc := tc
-		t.Run(tc.name, func(t *testing.T) {
-			err := m.validateElements(tc.elements)
-			if tc.expErr {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestModuleInstance_applyElements(t *testing.T) {
-	functionCounts := uint32(0xa)
-	m := &ModuleInstance{
-		Table:     &TableInstance{Table: make([]uintptr, 10)},
-		Functions: make([]*FunctionInstance, 10),
-		Engine:    &mockModuleEngine{},
-	}
-	targetIndex, targetOffset := uint32(1), byte(0)
-	targetIndex2, targetOffset2 := functionCounts-1, byte(0x8)
-	m.Functions[targetIndex] = &FunctionInstance{Type: &FunctionType{}, Index: Index(targetIndex)}
-	m.Functions[targetIndex2] = &FunctionInstance{Type: &FunctionType{}, Index: Index(targetIndex2)}
-	m.applyElements([]*ElementSegment{
-		{OffsetExpr: &ConstantExpression{Opcode: OpcodeI32Const, Data: []byte{targetOffset}}, Init: []uint32{targetIndex}},
-		{OffsetExpr: &ConstantExpression{Opcode: OpcodeI32Const, Data: []byte{targetOffset2}}, Init: []uint32{targetIndex2}},
-	})
-	require.Equal(t, uintptr(targetIndex), m.Table.Table[targetOffset])
-	require.Equal(t, uintptr(targetIndex2), m.Table.Table[targetOffset2])
 }
 
 func TestModuleInstance_decDependentCount(t *testing.T) {
